ml_service/requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
matplotlib
